requirements.txt
python-binance
requests
python-dotenv
pydantic
# -e.